returning R_xlen_t instead of int in many size() methods. #149

+++ inst/include/Rcpp/Dots.h
@@ -19,7 +19,7 @@ namespace Rcpp{
             }
         }
         
-        inline int size() const {
+        inline size_t size() const {
             return promises.size() ;    
         }
